organize client testdata into subfolders and normalize json file names
