+++ packages/client/test/sync/execution/vmexecution.spec.ts
@@ -5,9 +5,9 @@ import VM from '@ethereumjs/vm'
 import { Config } from '../../../lib/config'
 import { Chain } from '../../../lib/blockchain'
 import { VMExecution } from '../../../lib/execution'
-import blocksDataMainnet from './../../testdata/blocks_mainnet.json'
-import blocksDataGoerli from './../../testdata/blocks_goerli.json'
-import testnet from './../../testdata/testnet.json'
+import blocksDataMainnet from './../../testdata/blocks/mainnet.json'
+import blocksDataGoerli from './../../testdata/blocks/goerli.json'
+import testnet from './../../testdata/common/testnet.json'
 
 tape('[VMExecution]', async (t) => {
   t.test('Initialization', async (t) => {
