+++ packages/client/test/testdata/common/testnet.json
