+++ packages/client/test/testdata/geth-genesis/invalid-spurious-dragon.json
