+++ packages/client/test/testdata/geth-genesis/no-extra-data.json
