+++ packages/client/test/testdata/geth-genesis/pow.json
